requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
cryptography==36.0.2
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.74.1
h11==0.13.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.0
multidict==6.0.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
psycopg2-binary==2.9.3
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
uvicorn==0.17.5
yarl==1.7.2
